# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Daleel - AI-powered contract analysis platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /contracts:
    get:
      operationId: listContracts
      tags: [contracts]
      summary: List user contracts
      parameters:
        - name: contractType
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of contracts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContractSummary"
  /contracts/upload:
    post:
      operationId: uploadContract
      tags: [contracts]
      summary: Upload a contract file or paste text
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUploadInput"
      responses:
        "201":
          description: Contract uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{contractId}:
    get:
      operationId: getContract
      tags: [contracts]
      summary: Get a single contract with full analysis
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Contract details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateContract
      tags: [contracts]
      summary: Update contract metadata
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"

    delete:
      operationId: deleteContract
      tags: [contracts]
      summary: Delete a contract permanently
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /contracts/{contractId}/analyze:
    post:
      operationId: analyzeContract
      tags: [contracts]
      summary: Start AI analysis of an uploaded contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnalysisOptions"
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
  /contracts/{contractId}/export/{format}:
    get:
      operationId: exportContract
      tags: [contracts]
      summary: Export contract report (PDF or CSV)
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
        - name: format
          in: path
          required: true
          schema:
            type: string
            enum: [pdf, csv]
      responses:
        "200":
          description: Export file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /contracts/compare:
    post:
      operationId: compareContracts
      tags: [contracts]
      summary: Compare two contracts side by side
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompareInput"
      responses:
        "200":
          description: Comparison result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ComparisonResult"

  # ──────────────────────────────────────────────────────
  # CLAUSES
  # ──────────────────────────────────────────────────────
  /contracts/{contractId}/clauses:
    get:
      operationId: listClauses
      tags: [clauses]
      summary: List clauses for a contract
      parameters:
        - name: contractId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Clauses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Clause"

  # ──────────────────────────────────────────────────────
  # ASK DALEEL (Q&A)
  # ──────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH ──
...
    ContractUploadInput:
      type: object
      properties:
        pastedText:
          type: string
        fileBase64:
          type: string
        fileName:
          type: string
        fileType:
          type: string
        contractType:
          type: string
        language:
          type: string
          enum: [ar, en, mixed]
        title:
          type: string
    ContractUpdate:
      type: object
      properties:
        title:
          type: string
        contractType:
          type: string
    AnalysisOptions:
      type: object
      properties:
        explanationLevel:
          type: string
          enum: [simple, standard, detailed]
          default: standard
        outputLanguage:
          type: string
          enum: [ar, en]
          default: ar
    ContractSummary:
      type: object
      required: [id, title, contractType, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contractType:
          type: string
        status:
          type: string
          enum: [pending, processing, analyzed, failed]
        clarityScore:
          type: ["integer", "null"]
        highRiskCount:
          type: integer
        automaticRenewal:
          type: ["boolean", "null"]
        endDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    Contract:
      type: object
      required: [id, title, contractType, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        contractType:
          type: string
        originalFileName:
          type: ["string", "null"]
        language:
          type: ["string", "null"]
        extractedText:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, processing, analyzed, failed]
        clarityScore:
          type: ["integer", "null"]
        clarityExplanation:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
        summary:
          type: ["string", "null"]
        parties:
          type: array
          items:
            $ref: "#/components/schemas/Party"
        startDate:
          type: ["string", "null"]
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        duration:
          type: ["string", "null"]
        totalCost:
          type: ["number", "null"]
        currency:
          type: ["string", "null"]
        monthlyPayment:
          type: ["number", "null"]
        annualPayment:
          type: ["number", "null"]
        deposit:
          type: ["number", "null"]
        automaticRenewal:
          type: ["boolean", "null"]
        renewalNoticeDays:
          type: ["integer", "null"]
        cancellationAllowed:
          type: ["boolean", "null"]
        cancellationNoticeDays:
          type: ["integer", "null"]
        earlyCancellationPenalty:
          type: ["number", "null"]
        refundPolicy:
          type: ["string", "null"]
        trialPeriodDays:
          type: ["integer", "null"]
        clauses:
          type: array
          items:
            $ref: "#/components/schemas/Clause"
        financialDetails:
          type: array
          items:
            $ref: "#/components/schemas/FinancialDetail"
        contractDates:
          type: array
          items:
            $ref: "#/components/schemas/ContractDate"
        suggestedQuestions:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Party:
      type: object
      required: [id, name, role]
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
    Clause:
      type: object
      required: [id, category, title, simpleExplanation, riskLevel]
      properties:
        id:
          type: integer
        category:
          type: string
        title:
          type: string
        simpleExplanation:
          type: string
        riskLevel:
          type: string
          enum: [low, medium, high]
        sourcePage:
          type: ["integer", "null"]
        sourceText:
          type: ["string", "null"]
        confidence:
          type: ["number", "null"]
    FinancialDetail:
      type: object
      required: [id, type, amount, currency]
      properties:
        id:
          type: integer
        type:
          type: string
        amount:
          type: number
        currency:
          type: string
        description:
          type: ["string", "null"]
        sourcePage:
          type: ["integer", "null"]
        sourceText:
          type: ["string", "null"]
    ContractDate:
      type: object
      required: [id, type, date]
      properties:
        id:
          type: integer
        type:
          type: string
        date:
          type: string
          format: date
        description:
          type: ["string", "null"]
        sourcePage:
          type: ["integer", "null"]
    CompareInput:
      type: object
      required: [contractIdA, contractIdB]
      properties:
        contractIdA:
          type: integer
        contractIdB:
          type: integer
    ComparisonResult:
      type: object
      required: [contractA, contractB, aiSummary, differences]
      properties:
        contractA:
          $ref: "#/components/schemas/ContractSummary"
        contractB:
          $ref: "#/components/schemas/ContractSummary"
        aiSummary:
          type: string
        differences:
          type: array
          items:
            $ref: "#/components/schemas/ComparisonDifference"
    ComparisonDifference:
      type: object
      required: [field, labelAr, labelEn, valueA, valueB, advantage]
      properties:
        field:
          type: string
        labelAr:
          type: string
        labelEn:
          type: string
        valueA:
          type: string
        valueB:
          type: string
        advantage:
          type: string
          enum: [a, b, neutral]

    # ── Q&A ──
...