# lib/api-zod/src/generated/types/clause.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Daleel - AI-powered contract analysis platform
 * OpenAPI spec version: 0.1.0
 */
import type { ClauseRiskLevel } from './clauseRiskLevel';

export interface Clause {
  id: number;
  category: string;
  title: string;
  simpleExplanation: string;
  riskLevel: ClauseRiskLevel;
  /** @nullable */
  sourcePage?: number | null;
  /** @nullable */
  sourceText?: string | null;
  /** @nullable */
  confidence?: number | null;
}