# lib/api-zod/src/generated/types/party.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Daleel - AI-powered contract analysis platform
 * OpenAPI spec version: 0.1.0
 */

export interface Party {
  id: number;
  name: string;
  role: string;
}